fix: restructure folders
