+++ packages/components/typography/src/DisplayText/README.mdx
