+++ packages/components/typography/src/SectionHeading/README.mdx
