+++ packages/components/typography/src/SectionHeading/SectionHeading.spec.tsx
