+++ packages/components/typography/src/SectionHeading/SectionHeading.tsx
