+++ packages/components/typography/src/index.ts
@@ -1,14 +1,14 @@
-export { Heading } from './Heading';
-export type { HeadingProps, HeadingElement } from './Heading';
-export { Subheading } from './Subheading';
-export type { SubheadingProps } from './Subheading';
+export { Heading } from './Heading/Heading';
+export type { HeadingProps, HeadingElement } from './Heading/Heading';
+export { Subheading } from './Subheading/Subheading';
+export type { SubheadingProps } from './Subheading/Subheading';
 export { Typography } from './Typography';
 export type { TypographyProps } from './Typography';
-export { DisplayText } from './DisplayText';
-export type { DisplayTextProps } from './DisplayText';
-export { SectionHeading } from './SectionHeading';
-export type { SectionHeadingProps } from './SectionHeading';
-export { Paragraph } from './Paragraph';
-export type { ParagraphProps } from './Paragraph';
-export { Text } from './Text';
-export type { TextProps } from './Text';
+export { DisplayText } from './DisplayText/DisplayText';
+export type { DisplayTextProps } from './DisplayText/DisplayText';
+export { SectionHeading } from './SectionHeading/SectionHeading';
+export type { SectionHeadingProps } from './SectionHeading/SectionHeading';
+export { Paragraph } from './Paragraph/Paragraph';
+export type { ParagraphProps } from './Paragraph/Paragraph';
+export { Text } from './Text/Text';
+export type { TextProps } from './Text/Text';
