+++ packages/core/src/GlobalStyles/README.mdx
